python_api/requirements.txt
fastapi
uvicorn
redis
python-dotenv
requests
qdrant-client
numpy
websockets
langchain 
langchain-community 
langchain-core 
pypdf 
selenium 
unstructured
ollama
langchain-ollama
langchain-qdrant
streamlit
rq
psycopg2-binary
PyPDF2